+++ L1_Soc_Act.sh
@@ -13,23 +13,26 @@ cd $basedir
 OUTPUT=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Social_Act
 DATA=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/smoothing.feat/ICA_AROMA/denoised_func_data_nonaggr.nii.gz
 NVOLUMES=`fslnvols ${DATA}`
-rm -rf ${OUTPUT}.feat
+
+#remove if output file exists
+if [ -e ${OUTPUT}.feat ]; then
+  rm -rf ${OUTPUT}.feat
+fi
 
 #EV files
-EVMENTAL=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/mental.txt-
+EVMENTAL=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/mental.txt
 EVRND=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/rnd.txt
 
-#find and replace: run feat for smoothing
+#find and replace: run feat
 ITEMPLATE=${basedir}/templates/L1SocialAct.fsf
 OTEMPLATE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Social_Act.fsf
 sed -e 's@OUTPUT@'$OUTPUT'@g' \
 -e 's@DATA@'$DATA'@g' \
 -e 's@NVOLUMES@'$NVOLUMES'@g' \
--e 's@EVFEAR@'$EVFEAR'@g' \
--e 's@EVNEUT@'$EVNEUT'@g' \
+-e 's@EVMENTAL@'$EVMENTAL'@g' \
+-e 's@EVRND@'$EVRND'@g' \
 <$ITEMPLATE> $OTEMPLATE
-feat ${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Social_Act.fsf
-
+feat $OTEMPLATE
 
 #delete unused files
 #rm -rf ${OUTPUT}.feat/filtered_func_data.nii.gz
