+++ L1_WM_Act.sh
@@ -13,7 +13,11 @@ cd $basedir
 OUTPUT=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_WM_Act
 DATA=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/smoothing.feat/ICA_AROMA/denoised_func_data_nonaggr.nii.gz
 NVOLUMES=`fslnvols ${DATA}`
-rm -rf ${OUTPUT}.feat
+
+#remove if output file exists
+if [ -e ${OUTPUT}.feat ]; then
+  rm -rf ${OUTPUT}.feat
+fi
 
 #EV files
 EVDIR=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs
@@ -29,8 +33,9 @@ TBKPLACE=${EVDIR}/2bk_places.txt
 TBKTOOL=${EVDIR}/2bk_tools.txt
 
 #find and replace: run feat for smoothing
-ITEMPLATE=${basedir}/templates/L1_WM_Act.fsf
+ITEMPLATE=${basedir}/templates/L1WMAct.fsf
 OTEMPLATE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_WM_Act.fsf
+
 sed -e 's@OUTPUT@'$OUTPUT'@g' \
 -e 's@DATA@'$DATA'@g' \
 -e 's@NVOLUMES@'$NVOLUMES'@g' \
